requirements.txt
aiosqlite==0.20.0
alembic==1.13.1
apscheduler==3.10.4
asyncpg==0.29.0
durable_rules==2.0.4
fastapi-admin==1.0.5
fastapi-users[sqlalchemy]==12.1.2
fastapi==0.110.2
httpx==0.27.0
jinja2==3.1.3
passlib[bcrypt]==1.7.4
pydantic==2.7.1
pytest==8.2.0
python-dotenv==1.0.1
python-jose==3.3.0
redis==5.0.4
sentencepiece==0.2.0
sqlalchemy==2.0.29
torch>=2.2.0
transformers==4.40.1
uvicorn==0.29.0